pyproject.toml
[project]
name = "openfwi-src"
version = "0.1.0"
description = "Attempt at Full Waveform Inversion problem through Machine Learning with JAX!"
readme = "README.md"
authors = ["Ross John <RossDJohn14@gmail.com>"]
requires-python = ">=3.12"
dependencies = [
    "clu>=0.0.12",
    "datasets>=3.6.0",
    "fastapi>=0.111.1",
    "uvicorn[standard]>=0.30.3",
    "flax>=0.10.6",
    "jax>=0.6.1",
    "joblib>=1.5.1",
    "orbax-checkpoint>=0.11.13",
    "pyarrow>=20.0.0",
    "pytest>=8.4.0",
    "pydantic-settings>=2.10.1",
    "slowapi",
    "fastapi-cache2[redis]",
    "redis",
    "python-json-logger",
    "starlette-prometheus",
    "httpx",
    "pytest-asyncio>=1.0.0",
]
